switch the data type of Piece.Value to a byte data type instead of a full 4bytes int since the value will never exceed 9,this is 128 bytes in the Span<PieceInfo> ,now it fits entirely in an L1 cache , which will help making check detection fast with GC pressure and no allocation

+++ Assets/Scripts/Classes/Piece.cs
@@ -37,7 +37,7 @@ namespace Assets.Scripts.Classes
         /// The relative material worth of the piece used for AI evaluation 
         /// and game-state weighting.
         /// </summary>
-        public abstract uint Value { get; }
+        public abstract byte Value { get; }
 
         /// <summary>
         /// Evaluates the board state and populates the legal move-set for the piece.
